+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - #381 - fix the extra space in all `textarea` inputs
  - #378 - update help text for field `url` in styles: `link` and `button`
  - #383 - delete `reminder` only the data trigger is `finished` and not `started`
+ - `card` - properly use `keyword` in field `url_edit`
  
 
 # v6.7.0