+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - remove `toggle_switch` field from style `input` after creation of ne style `checkbox`. Now the new style contains the field `toggle_switch`
  - add `map_fields` in `data_config` when a `JSON` is returned.
  - add fields `color_background`, `color_border` and `color_text` in style `div`. They can be used to set the color for the div and its children. The fields can be set dynamically with `interpolation` input. Ex.: {{var_name}}
+ - add field `load_as_table` in style `entryList` - If enabled, the children are loaded inside a table.
 
 ### Bugfix
  - #379 - properly escape the naming pattern