+++ CHANGELOG.md
@@ -8,7 +8,7 @@
  - add fields `color_background`, `color_border` and `color_text` in style `div`. They can be used to set the color for the div and its children. The fields can be set dynamically with `interpolation` input. Ex.: {{var_name}}
  - add field `load_as_table` in style `entryList` - If enabled, the children are loaded inside a table.
  - show `entry_record` in the debug
- - add parent data from `entryLists` when used in nested situations. It is added with prefix `p_`. Check the values with debug in the console.
+ - add parent data from `entryLists` and `loop` when used in nested situations. It is added with prefix `p_`. Check the values with debug in the console.
 
 ### Bugfix
  - #379 - properly escape the naming pattern
@@ -18,6 +18,7 @@
  - `card` - properly use `keyword` in field `url_edit`
  - `redirect_at_end` url for forms work properly based on #keywords
  - `ui-preferences` - improve the check for settings
+ - deleting `entry` with `DELETE_RECORD_ID` is possible only if the entry belongs to the user, or the `form`'s property `own_entries_only` is disabled
  
 
 # v6.7.0