+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v6.8.0 (Unpublished)
+# v6.8.0
 ### New features
  - when queuing jobs from actions for `INTERNAL_FORM`, send the record_id of the newly inserted record
  - add new style `checkbox`, deprecate  the `checkbox` type in style `input`